requirements.txt
numpy>=1.26.0
pandas>=2.2.0
PyYAML>=6.0.1
tqdm>=4.66.0
opencv-python>=4.10.0.84
mediapipe>=0.10.14,<0.10.30
matplotlib>=3.8.0
scikit-learn>=1.5.0
torch>=2.3.0
torchvision>=0.18.0
kagglehub>=0.3.3